+++ crates/goose-mcp/src/computercontroller/mod.rs
@@ -1,11 +1,10 @@
 use base64::Engine;
 use etcetera::{choose_app_strategy, AppStrategy};
 use indoc::{formatdoc, indoc};
-use lopdf::{Document, Object, content::Content as PdfContent};
 use reqwest::{Client, Url};
 use serde_json::{json, Value};
 use std::{
-    collections::HashMap, fs, future::Future, io::Write, path::PathBuf, pin::Pin, sync::Arc,
+    collections::HashMap, fs, future::Future, path::PathBuf, pin::Pin, sync::Arc,
     sync::Mutex,
 };
 use tokio::process::Command;
@@ -20,6 +19,8 @@ use mcp_core::{
 };
 use mcp_server::router::CapabilitiesBuilder;
 use mcp_server::Router;
+
+mod pdf_tool;
 
 mod platform;
 use platform::{create_system_automation, SystemAutomation};
@@ -692,179 +693,7 @@ impl ComputerControllerRouter {
             .and_then(|v| v.as_str())
             .ok_or_else(|| ToolError::InvalidParameters("Missing 'operation' parameter".into()))?;
 
-        // Open and parse the PDF file
-        let doc = Document::load(path)
-            .map_err(|e| ToolError::ExecutionError(format!("Failed to open PDF file: {}", e)))?;
-
-        let result = match operation {
-            "extract_text" => {
-                let mut text = String::new();
-
-                // Iterate over each page in the document
-                for (page_num, page_id) in doc.get_pages() {
-                    text.push_str(&format!("Page {}:\n", page_num));
-
-                    // Try to get text from page contents
-                    if let Ok(page_obj) = doc.get_object(page_id) {
-                        if let Ok(page_dict) = page_obj.as_dict() {
-                            // Try to get text from Contents stream
-                            if let Ok(contents) = page_dict.get(b"Contents").and_then(|c| c.as_reference()) {
-                                if let Ok(content_obj) = doc.get_object(contents) {
-                                    if let Ok(stream) = content_obj.as_stream() {
-                                        if let Ok(content_data) = stream.get_plain_content() {
-                                            if let Ok(content) = PdfContent::decode(&content_data) {
-                                                // Process each operation in the content stream
-                                                for operation in content.operations {
-                                                    match operation.operator.as_ref() {
-                                                        // "Tj" operator: show text
-                                                        "Tj" => {
-                                                            for operand in operation.operands {
-                                                                if let Object::String(ref bytes, _) = operand {
-                                                                    if let Ok(s) = std::str::from_utf8(bytes) {
-                                                                        text.push_str(s);
-                                                                    }
-                                                                }
-                                                            }
-                                                            text.push(' ');
-                                                        },
-                                                        // "TJ" operator: show text with positioning
-                                                        "TJ" => {
-                                                            if let Some(Object::Array(ref arr)) = operation.operands.get(0) {
-                                                                let mut last_was_text = false;
-                                                                for element in arr {
-                                                                    match element {
-                                                                        Object::String(ref bytes, _) => {
-                                                                            if let Ok(s) = std::str::from_utf8(bytes) {
-                                                                                if last_was_text {
-                                                                                    text.push(' ');
-                                                                                }
-                                                                                text.push_str(s);
-                                                                                last_was_text = true;
-                                                                            }
-                                                                        },
-                                                                        Object::Integer(offset) => {
-                                                                            // Large negative offsets often indicate word spacing
-                                                                            if *offset < -100 {
-                                                                                text.push(' ');
-                                                                                last_was_text = false;
-                                                                            }
-                                                                        },
-                                                                        Object::Real(offset) => {
-                                                                            if *offset < -100.0 {
-                                                                                text.push(' ');
-                                                                                last_was_text = false;
-                                                                            }
-                                                                        },
-                                                                        _ => {}
-                                                                    }
-                                                                }
-                                                                text.push(' ');
-                                                            }
-                                                        },
-                                                        _ => () // Ignore other operators
-                                                    }
-                                                }
-                                            }
-                                        }
-                                    }
-                                }
-                            }
-                        }
-                    }
-                    text.push('\n');
-                }
-
-                if text.trim().is_empty() {
-                    format!("No text found in PDF")
-                } else {
-                    format!("Extracted text from PDF:\n\n{}", text)
-                }
-            }
-
-            "extract_images" => {
-                let cache_dir = self.get_cache_path("pdf_images", "");
-                fs::create_dir_all(&cache_dir).map_err(|e| {
-                    ToolError::ExecutionError(format!(
-                        "Failed to create image cache directory: {}",
-                        e
-                    ))
-                })?;
-
-                let mut images = Vec::new();
-                let mut image_count = 0;
-
-                for (_, page_id) in doc.get_pages() {
-                    if let Ok(page_obj) = doc.get_object(page_id) {
-                        if let Ok(page_dict) = page_obj.as_dict() {
-                            if let Ok(resources) =
-                                page_dict.get(b"Resources").and_then(|r| r.as_dict())
-                            {
-                                if let Ok(xobjects) =
-                                    resources.get(b"XObject").and_then(|x| x.as_dict())
-                                {
-                                    for (_, xobject) in xobjects.iter() {
-                                        if let Ok(xobject_id) = xobject.as_reference() {
-                                            if let Ok(xobject) = doc.get_object(xobject_id) {
-                                                if let Ok(stream) = xobject.as_stream() {
-                                                    // Check if it's an image by looking at the Subtype
-                                                    if let Ok(subtype) = stream
-                                                        .dict
-                                                        .get(b"Subtype")
-                                                        .and_then(|s| s.as_name())
-                                                    {
-                                                        if subtype == b"Image" {
-                                                            // Get the raw image data
-                                                            if let Ok(data) =
-                                                                stream.get_plain_content()
-                                                            {
-                                                                let image_path =
-                                                                    cache_dir.join(format!(
-                                                                        "image_{}.jpg",
-                                                                        image_count
-                                                                    ));
-                                                                if let Ok(mut file) =
-                                                                    std::fs::File::create(
-                                                                        &image_path,
-                                                                    )
-                                                                {
-                                                                    if file.write_all(&data).is_ok()
-                                                                    {
-                                                                        images.push(format!(
-                                                                            "Saved image to: {}",
-                                                                            image_path.display()
-                                                                        ));
-                                                                        image_count += 1;
-                                                                    }
-                                                                }
-                                                            }
-                                                        }
-                                                    }
-                                                }
-                                            }
-                                        }
-                                    }
-                                }
-                            }
-                        }
-                    }
-                }
-
-                if images.is_empty() {
-                    format!("No images found in PDF")
-                } else {
-                    format!("Extracted images:\n{}", images.join("\n"))
-                }
-            }
-
-            _ => {
-                return Err(ToolError::InvalidParameters(format!(
-                    "Invalid operation: {}. Valid operations are: 'extract_text', 'extract_images'",
-                    operation
-                )))
-            }
-        };
-
-        Ok(vec![Content::text(result)])
+        crate::computercontroller::pdf_tool::pdf_tool(path, operation, &self.cache_dir).await
     }
 
     async fn cache(&self, params: Value) -> Result<Vec<Content>, ToolError> {
@@ -1054,102 +883,4 @@ impl Router for ComputerControllerRouter {
     }
 }
 
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use serde_json::json;
-    use std::path::PathBuf;
 
-    #[tokio::test]
-    async fn test_pdf_text_extraction() {
-        let router = ComputerControllerRouter::new();
-        let test_pdf_path = PathBuf::from(env!("CARGO_MANIFEST_DIR"))
-            .join("src/computercontroller/tests/data/test.pdf");
-
-        let result = router
-            .pdf_tool(json!({
-                "path": test_pdf_path.to_str().unwrap(),
-                "operation": "extract_text"
-            }))
-            .await;
-
-        assert!(result.is_ok(), "PDF text extraction should succeed");
-        let content = result.unwrap();
-        assert!(!content.is_empty(), "Extracted text should not be empty");
-        let text = content[0].as_text().unwrap();
-        println!("Extracted text: {}", text);
-        assert!(text.contains("Page 1"), "Should contain page marker");
-    }
-
-    #[tokio::test]
-    async fn test_pdf_image_extraction() {
-        let router = ComputerControllerRouter::new();
-        let test_pdf_path = PathBuf::from(env!("CARGO_MANIFEST_DIR"))
-            .join("src/computercontroller/tests/data/test_image.pdf");
-
-        // First try text extraction to see the content
-        let text_result = router
-            .pdf_tool(json!({
-                "path": test_pdf_path.to_str().unwrap(),
-                "operation": "extract_text"
-            }))
-            .await;
-
-        println!(
-            "Text content: {}",
-            text_result.unwrap()[0].as_text().unwrap()
-        );
-
-        // Now try image extraction
-        let result = router
-            .pdf_tool(json!({
-                "path": test_pdf_path.to_str().unwrap(),
-                "operation": "extract_images"
-            }))
-            .await;
-
-        assert!(result.is_ok(), "PDF image extraction should succeed");
-        let content = result.unwrap();
-        assert!(
-            !content.is_empty(),
-            "Image extraction result should not be empty"
-        );
-        let text = content[0].as_text().unwrap();
-        println!("Image extraction result: {}", text);
-
-        // Should either find images or explicitly state none were found
-        assert!(
-            text.contains("Saved image to:") || text.contains("No images found"),
-            "Should either save images or report none found"
-        );
-    }
-
-    #[tokio::test]
-    async fn test_pdf_invalid_path() {
-        let router = ComputerControllerRouter::new();
-        let result = router
-            .pdf_tool(json!({
-                "path": "nonexistent.pdf",
-                "operation": "extract_text"
-            }))
-            .await;
-
-        assert!(result.is_err(), "Should fail with invalid path");
-    }
-
-    #[tokio::test]
-    async fn test_pdf_invalid_operation() {
-        let router = ComputerControllerRouter::new();
-        let test_pdf_path = PathBuf::from(env!("CARGO_MANIFEST_DIR"))
-            .join("src/computercontroller/tests/data/test.pdf");
-
-        let result = router
-            .pdf_tool(json!({
-                "path": test_pdf_path.to_str().unwrap(),
-                "operation": "invalid_operation"
-            }))
-            .await;
-
-        assert!(result.is_err(), "Should fail with invalid operation");
-    }
-}
